requirements.txt
argon2-cffi==20.1.0
boto3==1.17.69
botocore==1.20.112
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers
django-crispy-forms==1.10.0
django-crum==0.7.9
django-environ==0.4.5
django-extensions==3.1.2
django-filter==2.4.0
django-resized==0.3.11
django-simple-history==2.12.0
django-storages==1.11.1
djangorestframework==3.13.1
drf-writable-nested==0.6.3
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
jmespath==0.10.0
lml==0.1.0
Markdown==3.3.6
MarkupSafe
openpyxl==3.0.9
packaging==21.3
Pillow==8.2.0
psycopg2-binary==2.9.1
pycparser==2.21
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.4
pytz==2020.4
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
telesign==2.2.2
twilio==6.52.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp
Celery
